- Progressing update to Ubuntu 24.04 - TODO: Resolve tc:: ambiguity --> Remove terminal_utils.hpp from ulisse_msgs and leave only Futils.h in ctrl_toolbox

+++ ulisse_ctrl/src/ulisse_ctrl/kinematic_vehicle_controller.cpp
@@ -571,7 +571,7 @@ void VehicleController::NavFilterCB(const ulisse_msgs::msg::NavFilterData::Share
     Eigen::Vector3d bodyPosUTM;
     ctb::LatLong2LocalUTM(ctrlData_->inertialF_linearPosition, 0.0, centroidLocation_, bodyPosUTM);
 
-    Eigen::Vector3d worldF_vehicleLinearPosition(bodyPosUTM.x, bodyPosUTM.y, 0.0);
+    Eigen::Vector3d worldF_vehicleLinearPosition(static_cast<double>(bodyPosUTM.x()), static_cast<double>(bodyPosUTM.y()), 0.0);
 
     // Updating the robot model
     Eigen::TransformationMatrix worldF_T_vehicleF;
